+++ client/app/account/index.js
@@ -5,10 +5,10 @@ import uiRouter from 'angular-ui-router';
 import routes from './account.routes';
 import login from './login';
 import settings from './settings';
-import signup from './signup';
+//import signup from './signup';
 import oauthButtons from '../../components/oauth-buttons';
 
-export default angular.module('shyApp.account', [uiRouter, login, settings, signup, oauthButtons])
+export default angular.module('shyApp.account', [uiRouter, login, settings, /*signup,*/ oauthButtons])
   .config(routes)
   .run($rootScope => {
     'ngInject';