+++ client/app/account/signup/index.js
@@ -3,6 +3,6 @@
 import angular from 'angular';
 import SignupController from './signup.controller';
 
-export default angular.module('shyApp.signup', [])
+export default angular.module('shyApp.signupOld', [])
   .controller('SignupController', SignupController)
   .name;