+++ client/app/app.js
@@ -41,6 +41,7 @@ import registerPage from './register/register.component';
 import privacyPage from './privacy/privacy.component';
 import termsPage from './terms/terms.component';
 import adminPage from './admin/admin.component';
+import signupPage from './signup/signup.component';
 
 // Filters
 import upcoming from '../components/upcoming/upcoming.filter';
@@ -55,14 +56,14 @@ import './app.scss';
 
 // Inject everything into shyApp
 angular.module('shyApp', [ngCookies, ngResource, ngMessages, ngSanitize, uiRouter, uiBootstrap, _Auth, account, adminPage, navbar, banner, footer,
-  mainPage, classesPage, workshopsPage, locationsPage, teachersPage, cartPage, confirmationPage, registerPage, privacyPage, termsPage, constants, util, upcoming,
+  mainPage, classesPage, workshopsPage, locationsPage, teachersPage, cartPage, confirmationPage, registerPage, privacyPage, termsPage, signupPage, constants, util, upcoming,
   htmlid, daytodate, tweet, CartModule, dirPagination, loadingBar, compareTo])
   .config(routeConfig)
   .run(($rootScope, $location, Auth) => {
     'ngInject';
     // Redirect to login if route requires auth and you're not logged in
     $rootScope.$on('$stateChangeStart', (event, next) => {
-      Auth.isLoggedIn(function(loggedIn) {
+      Auth.isLoggedIn(loggedIn => {
         if(next.authenticate && !loggedIn) {
           $location.path('/login');
         }