+++ client/app/cart/checkout.component.spec.js
@@ -1,10 +1,12 @@
 /* global describe, beforeEach, inject, it, expect */
 'use strict';
 import angular from 'angular';
+import checkoutPage from './checkout.component';
 import CartModule from '../../components/cartmodule/cart.module';
 
-describe('Component: CheckOutComponent', () => {
+describe('Component: CheckOutComponent', function() {
   // load the cart module
+  beforeEach(angular.mock.module(checkoutPage));
   beforeEach(angular.mock.module(CartModule));
 
   let CheckOutComponent;
@@ -14,7 +16,7 @@ describe('Component: CheckOutComponent', () => {
     CheckOutComponent = $componentController('checkout', {});
   }));
 
-  it('should ...', () => {
+  it('should ...', function() {
     expect(1).to.equal(1);
   });
 });